DONE: File refactors
