+++ tests/triple_barrier/unit/test_stop_loss_happy_path.py
@@ -1,5 +1,5 @@
 from dateutil import parser
-from triple_barrier.triple_barrier import StopLoss
+from triple_barrier.trade_labeler import StopLoss
 from triple_barrier.multi_barrier_types import TradeSide
