+++ tests/triple_barrier/unit/test_time_barrier_happy_path.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from dateutil import parser
-from triple_barrier.triple_barrier import (
+from triple_barrier.trade_labeler import (
                                            TimeBarrier,
                                            )
